Add Python script to check website for dead links
About a week ago Jameson Lopp reached out on Twitter to say that
checking for dead links was now taking up a large portion of his time.
This Python script should save a lot of time by automatically checking
every link in the website and then asking for confirmation by showing
the link in a browser. The dead links and their associated files will be
outputted to a broken_links.txt file.