Fix bugs in check_links.py script (#873)
The check_links.py script was still making stupid mistakes when it came
to filtering out the unwanted links. After spending over an hour trying
to figure it out and approaching the problem from multiple angles I
solved it in a stupidly simply way by just putting the links through the
filter multiple times. I am still unsure of why that works. Keynes was
mostly full of crap, but perhaps he was on to something with his "animal
spirits."

You also don't have to enter the path to the website directory anymore
as it now assumes that it is already in the website directory.

I was also able to remove two third party dependencies from the script
to make it only use two packages outside of the Python standard library.