Minor cleanup to check_links.py script (#884)
I had forgotten about the "w+" mode in Python when writing to a file
that creates the file if it does not exist. Previously the script was
using a try catch statement to achieve the same functionality. The
improvement makes the script smaller and less readable, which is far
more Pythonic.