Improve asynchronous downloading in link checker to avoid being blocked (#905)
* add gitignore with mac files

* Space out requests to repeated domains

The check_links.py script has an option to asynchronously download and
check all of the links. The problem with this is that we will be sending
multiple requests for different pages to the same domain, which clearly
identifies us as a bot (people generally don't download a dozen websites
at once). The fix is to sort the links by their top level domain and
space out the requests to ensure that a website is never visited more
than once in a three second time span.