add wasabi wallet to privacy software section (#917)
* add wasabi wallet to privacy software section

I recommend adding Wasabi Wallet to the Privacy Software section. Wasabi Wallet is a well-known, open-source, non-custodial Bitcoin wallet that prioritizes privacy. It is listed on Bitcoin.org and offers several unique features, such as CoinJoin and Tor integration, that make it a top choice for anyone looking to protect their privacy when using Bitcoin.

* changed balance lookup to CoinJoin